+++ packages/client-flows/src/sdk/instances.ts
@@ -137,9 +137,8 @@ export class Instances extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }