+++ packages/client-flows/src/sdk/triggers.ts
@@ -134,9 +134,8 @@ export class Triggers extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }
 
@@ -475,9 +474,8 @@ export class Triggers extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }