+++ packages/client-flows/src/sdk/workflows.ts
@@ -132,9 +132,8 @@ export class Workflows extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }